+++ src/components/molecules/activity-registration-form/ActivityRegistrationForm.tsx
@@ -4,7 +4,6 @@ import { Controller, useForm } from 'react-hook-form';
 import { useState } from 'react';
 import Select from '../select/Select';
 import Button from '@/components/atoms/button/Button';
-import { Calendar } from '@/components/ui/calendar';
 import ReservationTimePicker from '../reservation-time-picker/ReservationTimePicker';
 
 interface ActivitySettingData {
@@ -30,7 +29,10 @@ export default function ActivityRegistrationForm() {
 
   return (
     <form className="w-full" onSubmit={handleSubmit(onSubmit)}>
-      <Button text="등록하기" color="black" size="s" />
+      <div className="flex items-center justify-between">
+        <h1>내 체험 등록</h1>
+        <Button text="등록하기" color="black" size="s" type="submit" />
+      </div>
       <div className="flex flex-col gap-24pxr">
         <ActivitieSettingInput text="제목" />
         {/* 수정 예정입니다 */}
@@ -45,7 +47,7 @@ export default function ActivityRegistrationForm() {
         <ActivitieSettingInput text="가격" />
         <label>주소</label>
         <ActivitieSettingInput text="주소를 입력해주세요" />
-
+        <label>예약 가능한 시간대</label>
         <ReservationTimePicker />
         <label>배너 이미지</label>
         <ImageUploader title="banner" />
